reimplement parallelism on ipyparallel
